requirements.txt
faiss-cpu
numpy
pandas
sentence-transformers
lime
torch
tqdm
pypdf
transformers
scikit-learn
matplotlib
